flask-backend/requirements.txt
Flask==2.1.1
openai==0.27.6
pinecone-client
python-dotenv==0.19.2
numpy
flask-cors
flask_cors